requirements.txt
google-generativeai
gTTS
Pillow
python-dotenv
